# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Smart University Communication and Venue Notification System API",
    "version": "1.0.0",
    "description": "A comprehensive backend system for university communication, announcements, venue management, and real-time notifications",
    "contact": {
      "name": "University of Zimbabwe",
      "email": "support@uz.ac.zw"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/announcements": {
      "post": {
        "tags": ["Announcements"],
        "summary": "Create announcement (Lecturers/Admins)",
        "security": [{"bearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "content"],
                "properties": {
                  "title": {"type": "string"},
                  "content": {"type": "string"},
                  "targetAudience": {"type": "string", "enum": ["ALL", "STUDENTS", "LECTURERS"], "default": "ALL"},
                  "pinned": {"type": "boolean", "default": false}
                }
              }
            }
          }
        },
        "responses": {
          "201": {"description": "Announcement created"},
          "403": {"description": "Forbidden"}
        }
      },
      "get": {
        "tags": ["Announcements"],
        "summary": "Get all announcements",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "targetAudience", "in": "query", "schema": {"type": "string"}},
          {"name": "pinned", "in": "query", "schema": {"type": "boolean"}},
          {"name": "search", "in": "query", "schema": {"type": "string"}},
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "default": 10}}
        ],
        "responses": {
          "200": {"description": "Announcements retrieved"}
        }
      }
    },
    "/api/announcements/{id}": {
      "get": {
        "tags": ["Announcements"],
        "summary": "Get announcement by ID",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Announcement retrieved"},
          "404": {"description": "Not found"}
        }
      },
      "put": {
        "tags": ["Announcements"],
        "summary": "Update announcement (Author/Admin)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {"type": "string"},
                  "content": {"type": "string"},
                  "targetAudience": {"type": "string", "enum": ["ALL", "STUDENTS", "LECTURERS"]},
                  "pinned": {"type": "boolean"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {"description": "Announcement updated"},
          "403": {"description": "Forbidden"}
        }
      },
      "delete": {
        "tags": ["Announcements"],
        "summary": "Delete announcement (Author/Admin)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Announcement deleted"},
          "403": {"description": "Forbidden"}
        }
      }
    },
    "/api/announcements/{id}/comments": {
      "post": {
        "tags": ["Comments"],
        "summary": "Add comment to announcement",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["content"],
                "properties": {
                  "content": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {"description": "Comment added"},
          "404": {"description": "Announcement not found"}
        }
      }
    },
    "/api/comments/{id}": {
      "delete": {
        "tags": ["Comments"],
        "summary": "Delete comment (Author/Admin)",
        "security": [{"bearerAuth": []}],
        "parameters": [
          {"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}
        ],
        "responses": {
          "200": {"description": "Comment deleted"},
          "403": {"description": "Forbidden"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}